+++ server/src/sv_main.h
@@ -123,9 +123,12 @@ void SV_ClearPlayerQueue();
 
 void SV_UpdateSecretCount(player_t & player);
 void SV_UpdateMonsterRespawnCount();
-void SV_SendExecuteLineSpecial(byte special, line_t* line, AActor* activator, byte arg0, byte arg1, byte arg2, byte arg3, byte arg4);
-void SV_ACSExecuteSpecial(byte special, AActor* activator, const char* print, bool playerOnly, int arg0 = -1, int arg1 = -1, int arg2 = -1, int arg3 = -1,
-	int arg4 = -1, int arg5 = -1, int arg6 = -1, int arg7 = -1, int arg8 = -1);
+void SV_SendExecuteLineSpecial(byte special, line_t* line, AActor* activator, int arg0,
+                               int arg1, int arg2, int arg3, int arg4);
+void SV_ACSExecuteSpecial(byte special, AActor* activator, const char* print,
+                          bool playerOnly, int arg0 = -1, int arg1 = -1, int arg2 = -1,
+                          int arg3 = -1, int arg4 = -1, int arg5 = -1, int arg6 = -1,
+                          int arg7 = -1, int arg8 = -1);
 
 bool CompareQueuePosition(const player_t* p1, const player_t* p2);
